Minor auction eval fixes
Auc cancel could be used to crash an eval price, added new cancelled option to auctions that enables true when an auction is cancelled, which will stop the eval process.

Also a check for if a card was put up for above eval on auction and failed. If this is the case, base eval is input as the failed price instead of a lowered number.