Very Unstable/ Messy implementation of slash-commands and buttons
This is a WIP version of what will become a base implementation of slash commands.

Currently it's extremely messy and disjointed as I am attempting to learn how the system even works. Once I get a better grasp of it, a lot of this will be re-written to be cleaner and more user friendly/readable.